Renamed "settings" to "endpoints" in modules
Originally named settings after eve standart definition, renamed now to make purpose clearer
